ios: clean up remaining existing services
