+++ ios/PolkadotVault/Backend/NavigationServices/GeneralVerifierService.swift
