+++ ios/PolkadotVault/Backend/NavigationServices/PublicKeyDetailsService.swift
@@ -8,14 +8,11 @@
 import Foundation
 
 final class PublicKeyDetailsService {
-    private let navigation: NavigationCoordinator
     private let backendService: BackendService
 
     init(
-        navigation: NavigationCoordinator = NavigationCoordinator(),
         backendService: BackendService = BackendService()
     ) {
-        self.navigation = navigation
         self.backendService = backendService
     }
 
@@ -27,12 +24,5 @@ final class PublicKeyDetailsService {
         backendService.performCall({
             try removeDerivedKey(address: address, networkSpecsKey: networkSpecsKey)
         }, completion: completion)
-    }
-
-    func resetNavigationState(_ keyName: String, _ publicKeyDetails: String) {
-        navigation.performFake(navigation: .init(action: .start))
-        navigation.performFake(navigation: .init(action: .navbarKeys))
-        navigation.performFake(navigation: .init(action: .selectSeed, details: keyName))
-        navigation.performFake(navigation: .init(action: .selectKey, details: publicKeyDetails))
     }
 }
