+++ ios/PolkadotVault/Backend/NavigationServices/ScanTabService.swift
